backend/pyproject.toml
[project]
name = "voice-ai-runtime"
version = "0.1.0"
description = "Event-Driven Voice Runtime for call center AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "fasttext-wheel>=0.9.2",
    "httpx>=0.28.1",
    "msgspec>=0.20.0",
    "numpy>=2.4.2",
    "onnxruntime>=1.24.3",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "orjson>=3.11.7",
    "prometheus-client>=0.24.1",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "redis>=7.2.1",
    "sentence-transformers>=5.2.3",
    "sentry-sdk>=2.54.0",
    "sqlalchemy>=2.0.48",
    "structlog>=25.5.0",
    "uvicorn>=0.41.0",
    "uvloop>=0.22.1",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E", "F", "W",     # pycodestyle + pyflakes
    "I",                # isort
    "N",                # pep8-naming
    "UP",               # pyupgrade
    "ANN",              # flake8-annotations
    "ASYNC",            # flake8-async
    "S",                # flake8-bandit
    "B",                # flake8-bugbear
    "A",                # flake8-builtins
    "C4",               # flake8-comprehensions
    "RET",              # flake8-return
    "SIM",              # flake8-simplify
    "TCH",              # flake8-type-checking
]
ignore = [
    "ANN101",  # missing type annotation for self
    "ANN102",  # missing type annotation for cls
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ANN"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
fail_under = 85

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
    "testcontainers>=4.14.1",
]
